+++ src/kernel/chain.h
@@ -5,12 +5,31 @@
 #ifndef BITCOIN_KERNEL_CHAIN_H
 #define BITCOIN_KERNEL_CHAIN_H
 
-#include<iostream>
+#include <attributes.h>
+
+#include <iostream>
 
 class CBlock;
 class CBlockIndex;
+class CBlockUndo;
+class uint256;
+
 namespace interfaces {
-struct BlockInfo;
+//! Block data sent with blockConnected, blockDisconnected notifications.
+struct BlockInfo {
+    const uint256& hash;
+    const uint256* prev_hash = nullptr;
+    int height = -1;
+    int file_number = -1;
+    unsigned data_pos = 0;
+    const CBlock* data = nullptr;
+    const CBlockUndo* undo_data = nullptr;
+    // The maximum time in the chain up to and including this block.
+    // A timestamp that can only move forward.
+    unsigned int chain_time_max{0};
+
+    BlockInfo(const uint256& hash LIFETIMEBOUND) : hash(hash) {}
+};
 } // namespace interfaces
 
 namespace kernel {
